separar lol e comandos do arquivo bot.js
