+++ commands.json
@@ -1,0 +1,94 @@
+[
+  {
+    "name": "banido",
+    "description": "audio e gif do banido"
+  },
+  {
+    "name": "perdemo",
+    "description": "audio e gif do perdemo"
+  },
+  {
+    "name": "empatamo",
+    "description": "audio e gif do empatamo"
+  },
+  {
+    "name": "legal",
+    "description": "audio e imagem do legal"
+  },
+  {
+    "name": "machista",
+    "description": "finn machista"
+  },
+  {
+    "name": "triste",
+    "description": "eu fico mt triste com uma noticia dessas :("
+  },
+  {
+    "name": "confia",
+    "description": "só confia"
+  },
+  {
+    "name": "alek",
+    "description": "alek talking in english"
+  },
+  {
+    "name": "ohuhalek",
+    "description": "alek destruindo na pista de danca"
+  },
+  {
+    "name": "bota",
+    "description": "audio e gif do bota"
+  },
+  {
+    "name": "cavalo",
+    "description": "cavalo"
+  },
+  {
+    "name": "massachusetts",
+    "description": "massachusetts"
+  },
+  {
+    "name": "etsirt",
+    "description": "(: sassed aiciton amu moc etsirt tm ocif ue"
+  },
+  {
+    "name": "oi",
+    "description": "(: oi linda, ai kawaii"
+  },
+  {
+    "name": "nicolas",
+    "description": "(: turugudum cha cha turugu cha turuudum cha cha"
+  },
+  {
+    "name": "glub",
+    "description": "glub glub"
+  },
+  {
+    "name": "an",
+    "description": "ãn"
+  },
+  {
+    "name": "yamete",
+    "description": "yamete Kudasai"
+  },
+  {
+    "name": "fdp",
+    "description": "iiui"
+  },
+  {
+    "name": "ganhamo",
+    "description": "ganhamo"
+  },
+  {
+    "name": "derrota",
+    "description": "derrota"
+  },
+  {
+    "name": "eele",
+    "description": "caralho will smith?"
+  },
+  {
+    "name": "machista2",
+    "description": "vai dirigir a cozinha"
+  }
+]
